backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
google-generativeai==0.8.3
pillow==10.1.0
pytesseract==0.3.10
redis==5.0.1
aiofiles==23.2.1
httpx==0.25.2
aiohttp==3.9.1
